[#19206] YSQL: Capture complete execution time when query is retried
Summary:
In scenarios involving KConflict or KReadRestart, where retry loops are encountered, this modification aids in accurately calculating the total time taken by the query for execution rather than solely relying on the time of the last instance (which was the case before this diff). It's essential to note that pgss captures only the execution time and does not account for planning or optimisation times.
In a retry, there 2 phases: execution and exponential backoff, so the total time a query took can be represented as:
execution time is taken during retries + backoff time + successful attempts time.

This diff also changes the explain analyze execution time to match pgss entry.
So after this diff:
EXPLAIN ANALYZE planning time is only of the first try (not accumulation), execution time is the total time that the query takes (total time that the user waits).
PGSS execution time is the total time that the query takes (total time that the user waits).

Test Plan:
```
ybd --java-test 'org.yb.pgsql.TestYbRetryTime'
```

Reviewers: asaha, telgersma, kramanathan, pjain, sanketh, xCluster, hsunder, #yba-api-review!

Subscribers: bogdan, sanketh, asrivastava, yugaware, smishra, yql

Differential Revision: https://phorge.dev.yugabyte.com/D32585